@@ -4,8 +4,8 @@
   "description": "React dashboard card component",
   "main": "lib/main.js",
   "scripts": {
-    "test": "rackt test --ci --browsers Firefox",
-    "start": "node_modules/.bin/rackt dev"
+    "test": "rackt test --single-run --browsers Firefox",
+    "start": "rackt server"
   },
   "repository": {
     "type": "git",
@@ -21,11 +21,14 @@
   },
   "homepage": "https://github.com/instructure-react/react-dashcard",
   "devDependencies": {
-    "rackt-cli": "^0.3.2",
-    "react": "^0.13.3",
-    "react-flipcard": "^0.1.2"
+    "rackt-cli": "^0.5.2",
+    "react": "^0.14.0",
+    "react-addons-test-utils": "^0.14.0",
+    "react-dom": "^0.14.0",
+    "react-flipcard": "^0.2.0"
   },
   "peerDependencies": {
-    "react": "^0.13.3"
+    "react": "^0.14.0",
+    "react-dom": "^0.14.0"
   }
 }
