@@ -1,3 +1,3 @@
-module.exports = function (obj) {
+export default function(obj) {
   return JSON.parse(JSON.stringify(obj));
-};
+}
