@@ -1,7 +1,7 @@
 import React, {PropTypes} from 'react';
 import clone from '../helpers/clone';
 
-var STYLES = {
+const STYLES = {
   container: {
     position: 'absolute',
     top: 0,
@@ -25,7 +25,9 @@ export default React.createClass({
 
   propTypes: {
     background: PropTypes.string,
-    onSettingsClick: PropTypes.func.isRequired
+    onSettingsClick: PropTypes.func.isRequired,
+    settingsIcon: PropTypes.element,
+    children: PropTypes.any
   },
 
   getDefaultProps() {
@@ -39,7 +41,7 @@ export default React.createClass({
   },
 
   render() {
-    var containerStyle = clone(STYLES.container);
+    const containerStyle = clone(STYLES.container);
     containerStyle.backgroundColor = this.props.background;
 
     return (
@@ -49,7 +51,7 @@ export default React.createClass({
       >
         <a
           style={STYLES.settings}
-          href="javascript://"
+          href="#"
           onClick={this.handleSettingsClick}
           aria-label="Open color picker"
         >
