@@ -1,8 +1,8 @@
 import React, {PropTypes} from 'react';
 import clone from '../helpers/clone';
-var isSafari = navigator && /safari/i.test(navigator.userAgent);
+const isSafari = navigator && /safari/i.test(navigator.userAgent);
 
-var STYLES = {
+const STYLES = {
   container: {
     backgroundColor: 'rgba(0, 0, 0, 0.8)',
     position: 'absolute',
@@ -85,7 +85,7 @@ var STYLES = {
     'MozBorderRadiusTopright': 5,
     'MozBorderRadiusBottomright': 5,
     'borderTopRightRadius': 5,
-    'borderBottomRightRadius': 5,
+    'borderBottomRightRadius': 5
   },
   screenreaderOnly: {
     overflow: 'hidden',
@@ -97,22 +97,22 @@ var STYLES = {
   }
 };
 
-var COLORS = [
-  {hex:'#EF4437', name: 'Red'},
-  {hex:'#E71F63', name: 'Pink'},
-  {hex:'#8F3E97', name: 'Purple'},
-  {hex:'#65499D', name: 'Deep Purple'},
-  {hex:'#4554A4', name: 'Indigo'},
-  {hex:'#2083C5', name: 'Blue'},
-  {hex:'#35A4DC', name: 'Light Blue'},
-  {hex:'#09BCD3', name: 'Cyan'},
-  {hex:'#009688', name: 'Teal'},
-  {hex:'#43A047', name: 'Green'},
-  {hex:'#8BC34A', name: 'Light Green'},
-  {hex:'#FDC010', name: 'Yellow'},
-  {hex:'#F8971C', name: 'Orange'},
-  {hex:'#F0592B', name: 'Deep Orange'},
-  {hex:'#F06291', name: 'Light Pink'}
+const COLORS = [
+  {hex: '#EF4437', name: 'Red'},
+  {hex: '#E71F63', name: 'Pink'},
+  {hex: '#8F3E97', name: 'Purple'},
+  {hex: '#65499D', name: 'Deep Purple'},
+  {hex: '#4554A4', name: 'Indigo'},
+  {hex: '#2083C5', name: 'Blue'},
+  {hex: '#35A4DC', name: 'Light Blue'},
+  {hex: '#09BCD3', name: 'Cyan'},
+  {hex: '#009688', name: 'Teal'},
+  {hex: '#43A047', name: 'Green'},
+  {hex: '#8BC34A', name: 'Light Green'},
+  {hex: '#FDC010', name: 'Yellow'},
+  {hex: '#F8971C', name: 'Orange'},
+  {hex: '#F0592B', name: 'Deep Orange'},
+  {hex: '#F06291', name: 'Light Pink'}
 ];
 
 function randomId() {
@@ -124,9 +124,14 @@ export default React.createClass({
 
   propTypes: {
     value: PropTypes.string,
-    defaultColors: PropTypes.arrayOf(PropTypes.string),
+    defaultColors: PropTypes.arrayOf(PropTypes.shape({
+      hex: PropTypes.string,
+      name: PropTypes.string
+    })),
     onValueChange: PropTypes.func.isRequired,
-    onCancelClick: PropTypes.func.isRequired
+    onCancelClick: PropTypes.func.isRequired,
+    closeIcon: PropTypes.element,
+    selectedIcon: PropTypes.element
   },
 
   getDefaultProps() {
@@ -145,22 +150,14 @@ export default React.createClass({
     };
   },
 
-  commit(value) {
-    this.props.onValueChange(value);
-  },
-
-  close() {
-    this.props.onCancelClick();
-  },
-
   handleTileClick(value) {
     this.commit(value);
     this.close();
   },
 
   handleTileKeyDown(value, event) {
-    if (event.keyCode === 13 /*enter*/ ||
-        event.keyCode === 32 /*spacebar*/) {
+    if (event.keyCode === 13 /* enter */ ||
+        event.keyCode === 32 /* spacebar */) {
       event.preventDefault();
       this.commit(value);
       this.close();
@@ -168,7 +165,7 @@ export default React.createClass({
   },
 
   handleApplyClick() {
-    this.commit(this.refs.input.getDOMNode().value);
+    this.commit(this.refs.input.value);
     this.close();
   },
 
@@ -177,7 +174,7 @@ export default React.createClass({
   },
 
   handleInputKeyUp() {
-    this.setState({ value: this.refs.input.getDOMNode().value });
+    this.setState({ value: this.refs.input.value });
   },
 
   getValue() {
@@ -185,10 +182,10 @@ export default React.createClass({
   },
 
   renderColorTiles() {
-    return this.props.defaultColors.map(function (color) {
-      var title = color.name + ' (' + color.hex + ')';
-      var selected = color.hex === this.getValue().toUpperCase();
-      var tileStyle = clone(STYLES.tile);
+    return this.props.defaultColors.map((color) => {
+      const title = color.name + ' (' + color.hex + ')';
+      const selected = color.hex === this.getValue().toUpperCase();
+      const tileStyle = clone(STYLES.tile);
       tileStyle.backgroundColor = color.hex;
 
       return (
@@ -210,13 +207,13 @@ export default React.createClass({
           </span>
         </button>
       );
-    }.bind(this));
+    });
   },
- 
+
   render() {
-    var previewStyle = clone(STYLES.colorPreview);
+    const previewStyle = clone(STYLES.colorPreview);
     previewStyle.backgroundColor = this.getValue();
-    
+
     return (
       <div
         style={STYLES.container}
@@ -224,7 +221,7 @@ export default React.createClass({
       >
         <a
           style={STYLES.cancel}
-          href="javascript://"
+          href="#"
           onClick={this.handleCancelClick}
           aria-label="Close color picker"
         >
@@ -265,5 +262,13 @@ export default React.createClass({
         </div>
       </div>
     );
+  },
+
+  commit(value) {
+    this.props.onValueChange(value);
+  },
+
+  close() {
+    this.props.onCancelClick();
   }
 });
