@@ -1,9 +1,10 @@
-import React, {PropTypes} from 'react';
+import React, { PropTypes } from 'react';
+import { findDOMNode } from 'react-dom';
 import FlipCard from 'react-flipcard';
 import CardDetails from './CardDetails';
 import ColorPicker from './ColorPicker';
 
-var STYLES = {
+const STYLES = {
   container: {
     position: 'relative'
   }
@@ -14,8 +15,15 @@ export default React.createClass({
 
   propTypes: {
     background: PropTypes.string,
-    colors: PropTypes.arrayOf(PropTypes.string),
-    onBackgroundChange: PropTypes.func
+    colors: PropTypes.arrayOf(PropTypes.shape({
+      hex: PropTypes.string,
+      name: PropTypes.string
+    })),
+    onBackgroundChange: PropTypes.func,
+    closeIcon: PropTypes.element,
+    selectedIcon: PropTypes.element,
+    settingsIcon: PropTypes.element,
+    children: PropTypes.any
   },
 
   getInitialState() {
@@ -38,7 +46,7 @@ export default React.createClass({
 
   handleOnFlip(flipped) {
     if (flipped) {
-      this.refs.colorpicker.getDOMNode().querySelector('a[aria-label]').focus();
+      findDOMNode(this.refs.colorpicker).querySelector('a[aria-label]').focus();
     }
   },
 
@@ -46,14 +54,6 @@ export default React.createClass({
     if (this.state.showColorPicker && e.keyCode === 27) {
       this.closeColorPicker();
     }
-  },
-
-  openColorPicker() {
-    this.setState({ showColorPicker: true });
-  },
-
-  closeColorPicker() {
-    this.setState({ showColorPicker: false });
   },
 
   render() {
@@ -89,5 +89,13 @@ export default React.createClass({
         </FlipCard>
       </div>
     );
+  },
+
+  openColorPicker() {
+    this.setState({ showColorPicker: true });
+  },
+
+  closeColorPicker() {
+    this.setState({ showColorPicker: false });
   }
 });
